requirements.txt
djangorestframework
django-filter
django-extensions
django-cors-headers
django-cors-headers
requests
coverage
pika
Faker